+++ datn_cozypot_spring_boot/src/main/java/com/example/datn_cozypot_spring_boot/dto/request/DatBanUpdateRequest.java
@@ -29,5 +29,5 @@ public class DatBanUpdateRequest {
 
         private Integer trangThaiPhieu;
 
-        private Double vatApDung;
+        private BigDecimal vatApDung;
 }
