+++ datn_cozypot_spring_boot/src/main/java/com/example/datn_cozypot_spring_boot/entity/HoaDonThanhToan.java
@@ -76,7 +76,7 @@ public class HoaDonThanhToan {
     private BigDecimal tongTienThanhToan;
 
     @Column(name = "vat_ap_dung", precision = 18)
-    private Float vatApDung;
+    private BigDecimal vatApDung;
 
     @Column(name = "diem_su_dung")
     private Integer diemSuDung;
